+++ examples/stressTest.tsx
@@ -47,10 +47,7 @@ function StressTestExample() {
                 }));
             }
             batch.store();
-            await Promise.all([
-                model.requestElementData(nodes),
-                model.requestLinks(),
-            ]);
+            await model.requestData();
             model.history.reset();
 
             const canvas = view.findAnyCanvas();